+++ lib/devise/async/backend/base.rb
@@ -6,8 +6,12 @@ module Devise
           raise NotImplementedError, "Any DeviseAssync::Backend subclass should implement `self.enqueue`."
         end
 
+        # Loads the resource record and sends the email.
+        #
+        # It uses `orm_adapter` API to fetch the record in order to enforce
+        # compatibility among diferent ORMs.
         def perform(method, resource_class, resource_id)
-          resource = resource_class.constantize.find(resource_id)
+          resource = resource_class.constantize.to_adapter.get!(resource_id)
           mailer_class.send(method, resource).deliver
         end