+++ test/devise/async/proxy_test.rb
@@ -5,7 +5,7 @@ module Devise
     describe "Proxy" do
       it "gets called by devise operations and proxy to worker" do
         user = create_user
-        Worker.expects(:enqueue).with(:confirmation_instructions, "User", user.to_param)
+        Worker.expects(:enqueue).with(:confirmation_instructions, "User", [user.id])
         user.send_confirmation_instructions
       end
     end